+++ src/main/java/fpt/com/universitymanagement/common/Constant.java
@@ -1,0 +1,8 @@
+package fpt.com.universitymanagement.common;
+
+public final class Constant {
+    private Constant() {
+    
+    }
+    public static final String ACCOUNT_CONTROLLER = "/v1/api/accounts";
+}
