--- src/main/java/fpt/com/universitymanagement/common/constant.java
@@ -1,4 +1,0 @@
-package fpt.com.universitymanagement.common;
-
-public class constant {
-}
