+++ src/main/java/fpt/com/universitymanagement/entity/Account.java
@@ -1,16 +1,20 @@
 package fpt.com.universitymanagement.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.util.UUID;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "accounts")
+@Table(name = "accounts", schema="public")
 public class Account extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    @Column(name = "id", updatable = false, nullable = false)
+    private UUID id;
     @Column(name = "user_name", nullable = false)
     private String userName;
     @Column(name = "password", nullable = false)
