+++ src/main/java/fpt/com/universitymanagement/entity/BaseEntity.java
@@ -8,22 +8,22 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
 import java.time.LocalDate;
-import java.util.UUID;
 
 @MappedSuperclass
 @Getter
 @Setter
 public class BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "id", updatable = false, nullable = false)
-    private UUID id;
-    
     @CreatedDate
     @Column(name = "created_at", updatable = false)
     private LocalDate createdAt;
     
     @LastModifiedDate
     @Column(name = "updated_at")
-    private LocalDate  updatedAt;
+    private LocalDate updatedAt;
+    
+    @Column(name = "created_by", updatable = false)
+    private String createdBy;
+    
+    @Column(name = "updated_by")
+    private String updatedBy;
 }
