+++ src/main/resources/application.properties
@@ -10,13 +10,3 @@ server.port=${PORT:8080}
 spring.devtools.restart.enabled=true
 spring.flyway.baseline-on-migrate=true
 spring.flyway.enabled=true
-spring.security:
-oauth2:
-client:
-registration:
-google:
-clientId: YOUR_GOOGLE_CLIENT_ID
-clientSecret: YOUR_GOOGLE_CLIENT_SECRET
-scope:
-- email
-- profile
